undo a couple earlier changes and change differently. I think it may be ok now.

+++ src/FontStashSharp/FontGlyph.cs
@@ -10,7 +10,7 @@ namespace FontStashSharp
 		public int Index;
 		public int Size;
 		public Rectangle Bounds;
-		public float XAdvance;
+		public int XAdvance;
 		public int XOffset;
 		public int YOffset;
